segment

+++ WeygoIPhone/Tabs/Home/Model/WGTitleItem.h
@@ -10,7 +10,7 @@
 
 @interface WGTitleItem : WGObject
 
-@property (nonatomic, assign) long id;
+@property (nonatomic, assign) long long id;
 
 @property (nonatomic, strong) NSString *name;
